requirements.txt
streamlit
Pillow
pymupdf
google-generativeai
python-dotenv
